+++ src/main/java/com/taxamo/client/model/SettlementDailyStatsSchema.java
@@ -21,44 +21,64 @@ package com.taxamo.client.model;
 */
 
 
-import java.math.BigDecimal;
-public class N {
+import java.lang.Integer;
+public class SettlementDailyStatsSchema {
   /* Date for stats in yyyy-MM-dd'T'hh:mm:ss'Z' format. */
   private String dayRaw = null;
-  /* Transaction count. */
-  private BigDecimal value = null;
-  /* Transaction status (C or N). */
-  private String status = null;
+  /* B2C transaction count. */
+  private Integer b2c = null;
+  /* B2B transaction count. */
+  private Integer b2b = null;
+  /* Total transaction count. */
+  private Integer count = null;
+  /* Untaxed transaction count. */
+  private Integer untaxed = null;
   /* Date for stats in yyyy-MM-dd format. */
   private String day = null;
   public String getDayRaw() {
     return dayRaw;
   }
-  public N setDayRaw(String dayRaw) {
+  public SettlementDailyStatsSchema setDayRaw(String dayRaw) {
     this.dayRaw = dayRaw;
     return this;
   }
 
-  public BigDecimal getValue() {
-    return value;
+  public Integer getB2c() {
+    return b2c;
   }
-  public N setValue(BigDecimal value) {
-    this.value = value;
+  public SettlementDailyStatsSchema setB2c(Integer b2c) {
+    this.b2c = b2c;
     return this;
   }
 
-  public String getStatus() {
-    return status;
+  public Integer getB2b() {
+    return b2b;
   }
-  public N setStatus(String status) {
-    this.status = status;
+  public SettlementDailyStatsSchema setB2b(Integer b2b) {
+    this.b2b = b2b;
+    return this;
+  }
+
+  public Integer getCount() {
+    return count;
+  }
+  public SettlementDailyStatsSchema setCount(Integer count) {
+    this.count = count;
+    return this;
+  }
+
+  public Integer getUntaxed() {
+    return untaxed;
+  }
+  public SettlementDailyStatsSchema setUntaxed(Integer untaxed) {
+    this.untaxed = untaxed;
     return this;
   }
 
   public String getDay() {
     return day;
   }
-  public N setDay(String day) {
+  public SettlementDailyStatsSchema setDay(String day) {
     this.day = day;
     return this;
   }
@@ -66,10 +86,12 @@ public class N {
   @Override
   public String toString()  {
     StringBuilder sb = new StringBuilder();
-    sb.append("class N {\n");
+    sb.append("class SettlementDailyStatsSchema {\n");
     sb.append("  dayRaw: ").append(dayRaw).append("\n");
-    sb.append("  value: ").append(value).append("\n");
-    sb.append("  status: ").append(status).append("\n");
+    sb.append("  b2c: ").append(b2c).append("\n");
+    sb.append("  b2b: ").append(b2b).append("\n");
+    sb.append("  count: ").append(count).append("\n");
+    sb.append("  untaxed: ").append(untaxed).append("\n");
     sb.append("  day: ").append(day).append("\n");
     sb.append("}\n");
     return sb.toString();