+++ src/main/java/com/taxamo/client/model/Summary.java
@@ -21,47 +21,48 @@ package com.taxamo.client.model;
 */
 
 
-import java.util.*;
-import com.taxamo.client.model.Report;
-public class GetSettlementOut {
-  /* Settlement report. */
-  private List<Report> report = new ArrayList<Report>();
-  /* Quarter start date in yyyy-MM-dd'T'hh:mm:ss'Z' format. */
-  private String startDate = null;
-  /* Quarter end date in yyyy-MM-dd'T'hh:mm:ss'Z' format. */
-  private String endDate = null;
+import java.math.BigDecimal;
+public class Summary {
+  /* Quarter that this summary applies to. */
+  private String quarter = null;
+  /* Tax amount due in this quarter. */
+  private BigDecimal taxAmount = null;
+  /* In which currency code the settlement was calculated. */
+  private String currencyCode = null;
   /* If the quarter isn't closed yet, tax amount is indicative, as we cannot determine FX rate or all transactions yet. */
   private Boolean indicative = null;
   /* Date of ECB FX rate used for conversions in yyyy-MM-dd'T'hh:mm:ss'Z' format. */
   private String fxRateDate = null;
-  public List<Report> getReport() {
-    return report;
+  /* Tax entity that the tax is due. */
+  private String taxEntityName = null;
+  public String getQuarter() {
+    return quarter;
   }
-  public GetSettlementOut setReport(List<Report> report) {
-    this.report = report;
+  public Summary setQuarter(String quarter) {
+    this.quarter = quarter;
     return this;
   }
 
-  public String getStartDate() {
-    return startDate;
+  public BigDecimal getTaxAmount() {
+    return taxAmount;
   }
-  public GetSettlementOut setStartDate(String startDate) {
-    this.startDate = startDate;
+  public Summary setTaxAmount(BigDecimal taxAmount) {
+    this.taxAmount = taxAmount;
     return this;
   }
 
-  public String getEndDate() {
-    return endDate;
+  public String getCurrencyCode() {
+    return currencyCode;
   }
-  public GetSettlementOut setEndDate(String endDate) {
-    this.endDate = endDate;
+  public Summary setCurrencyCode(String currencyCode) {
+    this.currencyCode = currencyCode;
     return this;
   }
 
   public Boolean getIndicative() {
     return indicative;
   }
-  public GetSettlementOut setIndicative(Boolean indicative) {
+  public Summary setIndicative(Boolean indicative) {
     this.indicative = indicative;
     return this;
   }
@@ -69,20 +70,29 @@ public class GetSettlementOut {
   public String getFxRateDate() {
     return fxRateDate;
   }
-  public GetSettlementOut setFxRateDate(String fxRateDate) {
+  public Summary setFxRateDate(String fxRateDate) {
     this.fxRateDate = fxRateDate;
+    return this;
+  }
+
+  public String getTaxEntityName() {
+    return taxEntityName;
+  }
+  public Summary setTaxEntityName(String taxEntityName) {
+    this.taxEntityName = taxEntityName;
     return this;
   }
 
   @Override
   public String toString()  {
     StringBuilder sb = new StringBuilder();
-    sb.append("class GetSettlementOut {\n");
-    sb.append("  report: ").append(report).append("\n");
-    sb.append("  startDate: ").append(startDate).append("\n");
-    sb.append("  endDate: ").append(endDate).append("\n");
+    sb.append("class Summary {\n");
+    sb.append("  quarter: ").append(quarter).append("\n");
+    sb.append("  taxAmount: ").append(taxAmount).append("\n");
+    sb.append("  currencyCode: ").append(currencyCode).append("\n");
     sb.append("  indicative: ").append(indicative).append("\n");
     sb.append("  fxRateDate: ").append(fxRateDate).append("\n");
+    sb.append("  taxEntityName: ").append(taxEntityName).append("\n");
     sb.append("}\n");
     return sb.toString();
   }