+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+=== 1.0.3
+ * informative mode for transaction lines
+ * new settlement operations
+ * new refund operations
+ * ability to query for tax supported countries only
+ * update regression tests
+
+
 === 1.0.1 
 
 Initial release